+++ packages/presets/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.9.0] - 2023-08-29
+
 ### Added
 
 - Added `Card` component.